% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIAT.noprac.R
\name{cleanIAT.noprac}
\alias{cleanIAT.noprac}
\title{Data analysis function: Cleans the IAT without Practice Blocks}
\usage{
cleanIAT.noprac(
  crit1,
  crit2,
  timeout.drop = TRUE,
  timeout.ms = 10000,
  fasttrial.drop = FALSE,
  fasttrial.ms = 400,
  fastprt.drop = TRUE,
  fastprt.percent = 0.1,
  fastprt.ms = 300,
  error.penalty = FALSE,
  error.penalty.ms = 600,
  inclusive.sd = TRUE
)
}
\description{
A variant of \code{cleanIAT()}. Inputs and outputs are identical, except that this function accepts only two blocks. This can be used when practice blocks are omitted from the IAT.
}
\examples{
\dontrun{
### Collapse  IAT critical blocks  down ####
clean <- cleanIAT.noprac(dat$compatible.crit, dat$incompatible.crit)
}
}
\seealso{
See the help for \code{cleanIAT()}.
}
