% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iatgen.R
\name{combineIATfourblocks}
\alias{combineIATfourblocks}
\title{Data analysis function: Collapses IAT permutations down prior to cleaning and analysis}
\usage{
combineIATfourblocks(name1, name2, name3, name4)
}
\arguments{
\item{name1}{A vector of responses representing a critical block (either congruent or incongruent) of trials for one of the four IAT permutations.}

\item{name2}{A vector of responses representing a critical block (either congruent or incongruent) of trials for another of the four IAT permutations.}

\item{name3}{A vector of responses representing a critical block (either congruent or incongruent) of trials for another of the four IAT permutations.}

\item{name4}{A vector of responses representing a critical block (either congruent or incongruent) of trials for another of the four IAT permutations.}
}
\value{
Returns a single vector of responses that contains all four permutations collapsed into one variable.
}
\description{
The first step after importing IAT data is collapsing the IAT data into two variables, typically labeled “congruent” and “incongruent.” The Qualtrics IAT runs (by default) four permutations of the IAT, counterbalancing left/right starting positions of both targets and categories. This ensures that Target A, Target B, the Positive category, and the Negative category all start equally often on the left and right sides of the screen (although research has consistently failed to find any left/right bias on the IAT; e.g., Greenwald et al., 1998; Nosek et al., 2005). Each of the four permutations includes two critical blocks, a “congruent“ critical block (A + Pos and B + Neg) and an “incongruent” critical block (A + Neg and B + Pos), which occur on blocks 4 and 7 (order depending on the permutation). Thus, whereas data analysis requires two variables with critical responses, the IAT output contains 8 variables (2 blocks per each of 4 permutations) that need collapsing. This is done with the \code{combineIATfourblocks} function. Scripts for this procedure are included with iatgen or can be seen in the examples below. This function is run twice, once to collapse the four congruent blocks down to one variable and once to collapse the four incongruent blocks down to one variable. This function ensures that if the IAT is skipped, the participant's data are scored as missing (facilitating easy data cleaning and accurate reporting). Note that if one of the input variables is empty (e.g., due to a typo specifying the variable names), the function will warn the user.
}
\examples{
\dontrun{
### Collapse  IAT data down ####
dat$congruent <- combineIATfourblocks(dat$Q4.RP4, dat$Q18.LP4, dat$Q14.RN7, dat$Q28.LN7)
dat$incongruent <- combineIATfourblocks(dat$Q7.RP7, dat$Q21.LP7, dat$Q11.RN4, dat$Q25.LN4)
}
}

