% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATalpha.R
\name{IATalpha}
\alias{IATalpha}
\title{Data analysis function: Cronbach's Alpha}
\usage{
IATalpha(data)
}
\arguments{
\item{data}{an object created by \code{cleanIAT()} representing a cleaned IAT.}
}
\value{
Returns a list with alphas based on the practice trials, the critical trials, and the entire IAT. Also retuns a matrix `diffscores` of the differences in latency for all matched trials (pooled practice and critical into one matrix)
}
\description{
An alternative way of assessing the internal consistency of the IAT is to use Cronbach's Alpha. This method, used by Schnabel, Asendorpf, & Greenwald (2008) creates pairs of reaction times from compatible/incompatible blocks, calculates their difference scores, and enters those into a Cronbach's alpha analysis.
}
\examples{
\dontrun{
### RELIABILITY ANALYSIS - ESTIMATE ONLY ###
IATalpha(clean)$alpha.total

### RELIABILITY ANALYSIS - ENTIRE RELIABILITY OUTPUT ###
IATalpha(clean)
}
}
\references{
Schnabel, K., Asendorpf, J. B., & Greenwald, A. G. (2008). Using Implicit Association Tests for the Assessment of Implicit Personality Self-Concept. In The SAGE Handbook of Personality Theory and Assessment: Volume 2—Personality Measurement and Testing (pp. 508–528). SAGE Publications Ltd. https://doi.org/10.4135/9781849200479.n24
}
