% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iatgen.R
\name{IATreliability}
\alias{IATreliability}
\title{Data analysis function: Estiamte reliability of IAT}
\usage{
IATreliability(data)
}
\arguments{
\item{block1}{An object created by \code{cleanIAT()} (or \code{cleanIAT.prac()}) representing the congruent or incongrent data after cleaning is complete.}

\item{block2}{An object created by \code{cleanIAT()} (or \code{cleanIAT.prac()}) representing the congruent or incongrent data after cleaning is complete.}
}
\value{
Returns reliability estimate.
}
\description{
One can easily estimate the IAT reliability by scoring the IAT separately based on odd and even trials and compute a split-half reliability. This is accomplished using \code{IATsplithalf()}, which sorts trials in order by type (positive, negative, target A, target B), takes alternating trials in order of presentation, scores and correlates the IAT, and applies a split-half spearman-brown correction (De Houwer & De Bruycker, 2007). This ensures an even distribution of targets and categories in odd/even trialsets (De Houwer & De Bruycker, 2007).
}
\examples{
\dontrun{
### RELIABILITY ANALYSIS - ESTIMATE ONLY ###
IATsplithalf(congruent.clean, incongruent.clean)$reliability

### RELIABILITY ANALYSIS - ENTIRE RELIABILITY OUTPUT ###
IATsplithalf(congruent.clean, incongruent.clean)
}
}

