% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iatgen.R
\name{combineIATtwoblocks}
\alias{combineIATtwoblocks}
\title{Data analysis function: Collapses IAT permutations down prior to cleaning and analysis (two-permutation version)}
\usage{
combineIATtwoblocks(name1, name2)
}
\arguments{
\item{name1}{A vector of responses representing a critical block (either congruent or incongruent) of trials for one of the IAT permutations.}

\item{name2}{A vector of responses representing a critical block (either congruent or incongruent) of trials for the other IAT permutation.}
}
\value{
Returns a single vector of responses that contains all four permutations collapsed into one variable.
}
\description{
This function is a variation of \code{combineIATfourblocks()} but using two permutations instead of four. Some users may opt to reduce the number of permutations of the IAT (e.g., affixing one target to the left side and the other to the right across participants) for various reasons (e.g., to match an in-lab configuration or for other reasons). This can be done with this function.
}
\examples{
\dontrun{
### Example with only Target A on the right and Target B on the left  ####
dat$congruent <- combineIATtwoblocks(dat$Q4.RP4, dat$Q14.RN7)
dat$incongruent <- combineIATtwoblocks(dat$Q7.RP7, dat$Q11.RN4)
}
}

