% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATreliability.R
\name{IATreliability}
\alias{IATreliability}
\title{Data analysis function: Estimate reliability of IAT}
\usage{
IATreliability(data, inclusive.sd = TRUE)
}
\arguments{
\item{data}{An object created by \code{cleanIAT()} representing a cleaned IAT.}
}
\value{
Returns reliability estimate and split-half correlation. In addition, D scores for odd and even trials can be viewed, as can the component practice and critical D scores.
}
\description{
One can easily estimate the IAT reliability by scoring the IAT separately based on odd and even trials and compute a split-half reliability. This is accomplished using \code{IATreliability()}, which sorts trials in order by type (positive, negative, target A, target B), takes alternating trials in order of presentation, scores and correlates the IAT, and applies a split-half spearman-brown correction (De Houwer & De Bruycker, 2007). This ensures an even distribution of targets and categories in odd/even trialsets.
}
\examples{
\dontrun{
### RELIABILITY ANALYSIS - ESTIMATE ONLY ###
IATreliability(clean)$reliability

### RELIABILITY ANALYSIS - ENTIRE RELIABILITY OUTPUT ###
IATreliability(clean)
}
}
\references{
De Houwer, J., & De Bruycker, E. (2007). The Implicit Association Test outperforms the extrinsic affective Simon task as an implicit measure of inter-individual differences in attitudes. \emph{British Journal of Social Psychology, 46}, 401–421. https://doi.org/10.1348/014466606X130346
}
