% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iatgen.R
\name{mergeIATblock}
\alias{mergeIATblock}
\title{Data analysis function: Appends trials from two blocks together for analysis as one large block}
\usage{
mergeIATblock(b1, b2)
}
\arguments{
\item{b1}{A vector of responses, one per participant, representing a block of trials.}

\item{b2}{A vector of responses, one per participant, representing a different block of trials.}
}
\value{
Returns a single vector of responses that includes trials from \code{b1} and \code{b2}.
}
\description{
Allows the user to append trials from multiple blocks together for analysis as a larger block. Accepts as input two vectors of raw data (e.g., practice and critical blocks) and returns one vector. This procedure is not currently used in data cleaning but may be useful for researchers wishing to examine different properties of IAT data.
}

